vlc-snap: introduce vlc-snap image
Previously the `vlc-ubuntu-focal` image was used to build snap.

In order to be able to update snap to core22, ubuntu focal can is no longer be
the base distribution for snap packaging.  This introduces a new docker image
`vlc-snap` which can be upgraded when the requirements for Snapcraft are
updated.